+++ src/internals/price_window.rs
@@ -37,7 +37,7 @@ impl PriceWindow<true> {
 
 impl<const SUM_OF_SQUARES: bool> PriceWindow<SUM_OF_SQUARES> {
     #[inline]
-    pub fn add(&mut self, ohlcv: &impl Ohlcv) {
+    pub fn add(&mut self, ohlcv: &impl Ohlcv) -> Price {
         let price = match self.bar_state.handle(ohlcv) {
             BarAction::Advance(price) => {
                 if let Some(old_price) = self.window.push(price) {
@@ -58,6 +58,8 @@ impl<const SUM_OF_SQUARES: bool> PriceWindow<SUM_OF_SQUARES> {
         if SUM_OF_SQUARES {
             self.sum_of_squares += price * price;
         }
+
+        price
     }
 
     #[inline]
@@ -69,6 +71,14 @@ impl<const SUM_OF_SQUARES: bool> PriceWindow<SUM_OF_SQUARES> {
     pub fn sum_of_squares(&self) -> Option<Price> {
         assert!(SUM_OF_SQUARES, "sum_of_squares requires PriceWindow<true>");
         self.is_ready().then_some(self.sum_of_squares)
+    }
+
+    #[inline]
+    pub fn fold<B, F>(&self, init: B, cb: F) -> B
+    where
+        F: FnMut(B, &Price) -> B,
+    {
+        self.window.fold(init, cb)
     }
 
     #[inline]